fixed first task && added solution for second task

+++ practice_1/first-task.py
@@ -7,21 +7,21 @@ import math
 
 class FigureArea:
     @staticmethod
-    def getSquareArea(a:int,b:int)->int:
+    def getSquareArea(a:float,b:float)->float:
         return a*b
     
     @staticmethod
-    def getTriangleArea(a:int, h:int)->int:
+    def getTriangleArea(a:float, h:float)->float:
         return (a*h)/2
     
     @staticmethod
-    def getCircleArea(r:int)->int:
+    def getCircleArea(r:float)->float:
         return math.pi*(r**2)
     
 
-circleArea:int = FigureArea.getCircleArea(2)
-triangleArea:int = FigureArea.getTriangleArea(5,2)
-squareArea:int = FigureArea.getSquareArea(10,12)
+circleArea:float = FigureArea.getCircleArea(2)
+triangleArea:float = FigureArea.getTriangleArea(5,2)
+squareArea:float = FigureArea.getSquareArea(10,12)
 
 figureAreas = {
     "площадь круга": circleArea,
